Add implementation to modifyOneQuery()
