+++ src/functions/searches/modify-one-query.ts
@@ -6,11 +6,39 @@
  * MIT license. See the LICENSE file for details.
  **************************************************************************/
 
+import { isNull, pick } from 'lodash';
+import { filter, first, map } from 'rxjs/operators';
 import { APIContext } from '~/functions/utils';
 import { ElementFilter, Query } from '~/models';
+import { makeSubscribeToOneQueryParsing, ValidatedQuery } from './subscribe-to-query-parsing';
 
 export const makeModifyOneQuery = (context: APIContext) => {
-	return async (query: Query, modifier: ElementFilter): Promise<Query> => {
-		return {} as any;
+	const subscribeToOneQueryParsing = makeSubscribeToOneQueryParsing(context);
+	let querySubP: ReturnType<typeof subscribeToOneQueryParsing> | null = null;
+
+	return async (query: Query, filters: Array<ElementFilter>): Promise<Query> => {
+		if (isNull(querySubP)) querySubP = subscribeToOneQueryParsing();
+		const querySub = await querySubP;
+		const id = SEARCH_SOCKET_ID_GENERATOR.generate();
+
+		const parsingP = querySub.received$
+			.pipe(
+				filter(msg => msg.id === id),
+				map(msg => pick(msg, ['isValid', 'error']) as ValidatedQuery),
+				first(),
+			)
+			.toPromise();
+		querySub.send({ id, query, filters });
+
+		const parsed = await parsingP;
+		return parsed.query;
 	};
 };
+
+const SEARCH_SOCKET_ID_GENERATOR = (() => {
+	let current = 0;
+
+	return {
+		generate: (): number => ++current,
+	};
+})();
