+++ src/functions/searches/validate-one-query.ts
@@ -8,12 +8,12 @@
 
 import { isNull, pick } from 'lodash';
 import { filter, first, map } from 'rxjs/operators';
-import { ElementFilter, Query, RawQuery } from '~/models';
-import { RawElementFilter, toElementFilter } from '~/models/search';
-import { APIContext, APISubscription, apiSubscriptionFromWebSocket, buildURL, WebSocket } from '../utils';
+import { Query } from '~/models';
+import { APIContext } from '../utils';
+import { makeSubscribeToOneQueryParsing, ValidatedQuery } from './subscribe-to-query-parsing';
 
 export const makeValidateOneQuery = (context: APIContext) => {
-	const subscribeToOneQueryValidation = makeSubscribeToOneQueryValidation(context);
+	const subscribeToOneQueryValidation = makeSubscribeToOneQueryParsing(context);
 	let querySubP: ReturnType<typeof subscribeToOneQueryValidation> | null = null;
 
 	return async (query: Query): Promise<ValidatedQuery> => {
@@ -40,147 +40,3 @@ const SEARCH_SOCKET_ID_GENERATOR = (() => {
 		generate: (): number => ++current,
 	};
 })();
-
-const makeSubscribeToOneQueryValidation = (context: APIContext) => {
-	const templatePath = '/api/ws/search';
-	const url = buildURL(templatePath, { ...context, protocol: 'ws' });
-
-	return async (): Promise<
-		APISubscription<ValidatedQuery & { id: number }, { id: number; query: Query; filters: Array<ElementFilter> }>
-	> => {
-		const socket = new WebSocket(url, context.authToken ?? undefined);
-		const rawSubscription = apiSubscriptionFromWebSocket<
-			RawQueryValidationMessageReceived,
-			RawQueryValidationMessageSent
-		>(socket);
-		rawSubscription.send({ Subs: ['PONG', 'parse', 'search', 'attach'] });
-
-		const received$ = rawSubscription.received$.pipe(
-			filter((msg): msg is RawQueryValidationDesiredMessageReceived => {
-				try {
-					return (msg as RawQueryValidationDesiredMessageReceived)?.type === 'parse';
-				} catch {
-					return false;
-				}
-			}),
-			map(msg => ({ ...toValidatedQuery(msg.data), id: msg.data.Sequence })),
-		);
-
-		const sent$ = rawSubscription.sent$.pipe(
-			filter((msg): msg is RawQueryValidationDesiredMessageSent => {
-				try {
-					return (msg as RawQueryValidationDesiredMessageSent)?.type === 'parse';
-				} catch {
-					return false;
-				}
-			}),
-			map(msg => ({
-				id: msg.data.Sequence,
-				query: msg.data.SearchString,
-				filters: (msg.data.Filters ?? []).map(toElementFilter),
-			})),
-		);
-
-		const received: Array<ValidatedQuery & { id: number }> = [];
-		const sent: Array<{ id: number; query: Query; filters: Array<ElementFilter> }> = [];
-
-		received$.subscribe(receivedMessage => received.push(receivedMessage));
-		sent$.subscribe(sentMessage => sent.push(sentMessage));
-
-		return {
-			close: () => rawSubscription.close(),
-			send: async msg => {
-				const rawMsg: RawQueryValidationMessageSent = {
-					type: 'parse',
-					data: {
-						SearchString: msg.query,
-						Sequence: msg.id,
-					},
-				};
-				if (msg.filters.length !== 0)
-					rawMsg.data.Filters = msg.filters.map(f => ({
-						Tag: f.tag,
-						Module: f.module,
-						Path: f.path,
-						Op: f.operation,
-						Value: f.value,
-					}));
-				await rawSubscription.send(rawMsg);
-			},
-			received$,
-			received,
-			sent$,
-			sent,
-		};
-	};
-};
-
-type RawQueryValidationDesiredMessageSent = {
-	type: 'parse';
-	data: {
-		SearchString: string;
-		Sequence: number;
-		/** Include the Filters field during data exploration (data explorer) */
-		Filters?: Array<RawElementFilter>;
-	};
-};
-type RawQueryValidationMessageSent =
-	| RawQueryValidationDesiredMessageSent
-	| { Subs: Array<'PONG' | 'parse' | 'search' | 'attach'> };
-
-type RawQueryValidationDesiredMessageReceived = { type: 'parse'; data: RawValidatedQuery };
-type RawQueryValidationMessageReceived = RawQueryValidationDesiredMessageReceived | { Resp: 'ACK' };
-
-interface RawValidatedQueryBase {
-	CollapsingIndex: number;
-	ModuleHints: Array<{
-		Name: string;
-		Condensing: boolean;
-		ProducedEVs: Array<string> | null;
-		ConsumedEVs: Array<string> | null;
-		ResourcesNeeded: Array<string> | null;
-	}> | null;
-	ModuleIndex: number;
-	ParsedQuery: RawQuery;
-	Sequence: number;
-	Tags: Array<string> | null;
-	TimeZoomDisabled: boolean;
-}
-
-interface RawApprovedValidatedQuery extends RawValidatedQueryBase {
-	GoodQuery: true;
-	RawQuery: RawQuery;
-	RenderModule: string;
-}
-
-interface RawRepprovedValidatedQuery extends RawValidatedQueryBase {
-	GoodQuery: false;
-	ParseError: string;
-}
-
-type RawValidatedQuery = RawApprovedValidatedQuery | RawRepprovedValidatedQuery;
-
-interface QueryError {
-	message: string;
-	module: number;
-}
-
-export type ValidatedQuery = { isValid: true; error: null } | { isValid: false; error: QueryError };
-
-const toValidatedQuery = (raw: RawValidatedQuery): ValidatedQuery => {
-	switch (raw.GoodQuery) {
-		case true:
-			return {
-				isValid: true,
-				error: null,
-			};
-		case false:
-			return {
-				isValid: false,
-				error: {
-					message: raw.ParseError ?? 'Unknown error',
-					module: raw.ModuleIndex,
-				},
-			};
-	}
-};
